Feature/file system cache (#692)
* First pass at caching file system.

* default-file-system -> slang-file-system
fix problem with location("build.linux") confusing windows build for now.

* Added CompressedResult
Fix problem in Result construction with it being unsigned

* Add support for Path simplification.

* Testing for Path::Simplify.

* Refactored CacheFileSystem - automatically handles ISlangFileSystem or ISlangFileSystemExt appropriately.
Removed WrapFileSystem - because wasn't possible to emulate some of the behavior if just loadFile is implemented.
Split out StringBlob - so that no need to convert between ISlangBlob and String repeatidly.

* Remove unwanted code in ~CompileRequest

+++ slang.h
@@ -535,7 +535,7 @@ extern "C"
     //! Get the result code for the facility
 #define SLANG_GET_RESULT_CODE(r)        ((int32_t)((r) & 0xffff))
 
-#define SLANG_MAKE_ERROR(fac, code)        ((((int32_t)(fac)) << 16) | ((int32_t)(code)) | 0x80000000)
+#define SLANG_MAKE_ERROR(fac, code)        ((((int32_t)(fac)) << 16) | ((int32_t)(code)) | int32_t(0x80000000))
 #define SLANG_MAKE_SUCCESS(fac, code)    ((((int32_t)(fac)) << 16) | ((int32_t)(code)))
 
     /*************************** Facilities ************************************/
